Send tile results to client as soon as they are available
Don't wait for all tiles to be retrieved before sending the results to the client. Instead call `saveHandler` as part of the iteration over the tile ids, so each tile's content is streamed back to the client as soon as it is available.

This will greatly improve performance for searches that return a large number of tiles.